showing concept to have LocalProxy be able to DI it's AppolicationContextManager in such a way that IF IN STATEFUL environment, it will always create a
